Update the ComputeArgumentName binding.
Update the implementation, remove all the error returns and provide
the runtime_error with expressive message in case of error.
Add document helper to each method and update the returned values
and input arguments.
Correct the naming of the variables and update most of them to lower
case letters.
Start to provide the pickling support to serialize and de-serialize
the object in Python.