Update the SimulatorModel binding.
Correctly use the custom destructor.
Update the implementation, remove all the error returns and provide
the runtime_error with expressive message in case of error.
Add document helper to each method and update the returned values
and input arguments.
Correct the naming of the variables and update most of them to lower
case letters.