Update the Model binding.
Refactor the binding and use the provided wrapper class. By using
the current wrapper class and refactoring the implementation we
remove the need for user requirement to explicitely edetrsoy the
object.
Update the implementation, remove all the error returns and provide
the runtime_error with expressive message in case of error.
Add document helper to each method and update the returned values
and input arguments.
Correct the naming of the variables and update most of them to lower
case letters.